moved event point to separate controller
